+++ tui/src/args.rs
@@ -1,10 +1,6 @@
 //! Parsing command-line arguments.
 
-use clap::{
-    App,
-    AppSettings::{AllowNegativeNumbers, ColoredHelp},
-    Arg, Error, ErrorKind, Result as ClapResult,
-};
+use clap::{app_from_crate, AppSettings, Arg, ErrorKind, Result};
 use rlifesrc_lib::{
     rules::NtLifeGen, Config, NewState, PolyWorld, SearchOrder, Symmetry, Transform,
 };
@@ -30,11 +26,8 @@ pub struct Args {
 
 impl Args {
     /// Parses the command-line arguments.
-    pub(crate) fn parse() -> ClapResult<Self> {
-        let mut app = App::new(env!("CARGO_PKG_NAME"))
-            .version(env!("CARGO_PKG_VERSION"))
-            .author(env!("CARGO_PKG_AUTHORS"))
-            .about(env!("CARGO_PKG_DESCRIPTION"))
+    pub(crate) fn parse() -> Result<Self> {
+        let mut app = app_from_crate!()
             .long_about(
                 "Searching for patterns in Conway's Game of Life\n\
                  \n\
@@ -45,9 +38,9 @@ impl Args {
                  https://github.com/AlephAlpha/rlifesrc/blob/master/tui/README.md (In Chinese)\n\
                  https://github.com/AlephAlpha/rlifesrc/blob/master/tui/README_en.md (In English)\n",
             )
-            .settings(&[AllowNegativeNumbers, ColoredHelp])
+            .global_setting(AppSettings::AllowNegativeNumbers)
             .arg(
-                Arg::with_name("CONFIG")
+                Arg::new("CONFIG")
                     .help("Read config from a file")
                     .long_help(
                         "Read config from a file\n\
@@ -55,65 +48,59 @@ impl Args {
                          When a config file is provided, all the other flags and options, \
                          except --all (-a), --reset-time, --no-tui (-n), are ignored.\n",
                     )
-                    .short("C")
+                    .short('C')
                     .long("config")
                     .takes_value(true),
             )
             .arg(
-                Arg::with_name("X")
+                Arg::new("X")
                     .help("Width of the pattern")
-                    .required_unless("CONFIG")
+                    .required_unless_present("CONFIG")
                     .index(1)
                     .validator(|x| {
-                        if is_positive(&x) {
-                            Ok(())
-                        } else {
-                            Err(String::from("width must be a positive integer"))
-                        }
+                        is_positive(x)
+                            .then(|| ())
+                            .ok_or_else(|| String::from("width must be a positive integer"))
                     }),
             )
             .arg(
-                Arg::with_name("Y")
+                Arg::new("Y")
                     .help("Height of the pattern")
-                    .required_unless("CONFIG")
+                    .required_unless_present("CONFIG")
                     .index(2)
                     .validator(|y| {
-                        if is_positive(&y) {
-                            Ok(())
-                        } else {
-                            Err(String::from("height must be a positive integer"))
-                        }
+                        is_positive(y)
+                            .then(|| ())
+                            .ok_or_else(|| String::from("height must be a positive integer"))
                     }),
             )
             .arg(
-                Arg::with_name("P")
+                Arg::new("P")
                     .help("Period of the pattern")
                     .default_value("1")
                     .index(3)
                     .validator(|p| {
-                        if is_positive(&p) {
-                            Ok(())
-                        } else {
-                            Err(String::from("period must be a positive integer"))
-                        }
+                        is_positive(p)
+                            .then(|| ())
+                            .ok_or_else(|| String::from("period must be a positive integer"))
                     }),
             )
             .arg(
-                Arg::with_name("DX")
+                Arg::new("DX")
                     .help("Horizontal translation")
                     .default_value("0")
                     .index(4)
                     .validator(|d| d.parse::<i32>().map(|_| ()).map_err(|e| e.to_string())),
             )
             .arg(
-                Arg::with_name("DY")
+                Arg::new("DY")
                     .help("Vertical translation")
                     .default_value("0")
                     .index(5)
                     .validator(|d| d.parse::<i32>().map(|_| ()).map_err(|e| e.to_string())),
             )
             .arg(
-                Arg::with_name("DIAG")
+                Arg::new("DIAG")
                     .help("Diagonal width")
                     .long_help(
                         "Diagonal width\n\
@@ -121,12 +108,12 @@ impl Args {
                          where abs(x - y) >= n are assumed to be dead.\n\
                          If this value is set to 0, it would be ignored.\n",
                     )
-                    .short("d")
+                    .short('d')
                     .long("diag")
                     .takes_value(true)
                     .default_value("0")
                     .validator(|d| {
-                        if is_positive(&d) || d == "0" {
+                        if is_positive(d) || d == "0" {
                             Ok(())
                         } else {
                             Err(String::from("diagonal width must be a positive integer"))
@@ -134,7 +121,7 @@ impl Args {
                     }),
             )
             .arg(
-                Arg::with_name("TRANSFORM")
+                Arg::new("TRANSFORM")
                     .help("Transformation of the pattern")
                     .long_help(
                         "Transformation of the pattern\n\
@@ -146,14 +133,14 @@ impl Args {
                          \"R\" means counterclockwise rotation.\n\
                          \"F\" means flipping (reflection) across an axis.\n",
                     )
-                    .short("t")
+                    .short('t')
                     .long("transform")
                     .takes_value(true)
                     .possible_values(&["Id", "R90", "R180", "R270", "F|", "F-", "F\\", "F/"])
                     .default_value("Id"),
             )
             .arg(
-                Arg::with_name("SYMMETRY")
+                Arg::new("SYMMETRY")
                     .help("Symmetry of the pattern")
                     .long_help(
                         "Symmetry of the pattern\n\
@@ -162,7 +149,7 @@ impl Args {
                          Logic Life Search.\n\
                          See [https://conwaylife.com/wiki/Symmetry] \n",
                     )
-                    .short("s")
+                    .short('s')
                     .long("symmetry")
                     .takes_value(true)
                     .possible_values(&[
@@ -171,14 +158,14 @@ impl Args {
                     .default_value("C1"),
             )
             .arg(
-                Arg::with_name("RULE")
+                Arg::new("RULE")
                     .help("Rule of the cellular automaton")
                     .long_help(
                         "Rule of the cellular automaton\n\
                          Supports Life-like, isotropic non-totalistic, hexagonal, MAP rules, \
                          and their corresponding Generations rules.\n",
                     )
-                    .short("r")
+                    .short('r')
                     .long("rule")
                     .takes_value(true)
                     .default_value("B3/S23")
@@ -189,13 +176,13 @@ impl Args {
                     }),
             )
             .arg(
-                Arg::with_name("ORDER")
+                Arg::new("ORDER")
                     .help("Search order")
                     .long_help(
                         "Search order\n\
                          Row first or column first.\n",
                     )
-                    .short("o")
+                    .short('o')
                     .long("order")
                     .takes_value(true)
                     .possible_values(&[
@@ -211,49 +198,49 @@ impl Args {
                     .default_value("automatic"),
             )
             .arg(
-                Arg::with_name("CHOOSE")
+                Arg::new("CHOOSE")
                     .help("How to choose a state for unknown cells\n")
-                    .short("c")
+                    .short('c')
                     .long("choose")
                     .takes_value(true)
                     .possible_values(&["dead", "alive", "random", "d", "a", "r"])
                     .default_value("alive"),
             )
             .arg(
-                Arg::with_name("MAX")
+                Arg::new("MAX")
                     .help("Upper bound of numbers of minimum living cells in all generations")
                     .long_help(
                         "Upper bound of numbers of minimum living cells in all generations\n\
                          If this value is set to 0, it means there is no limitation.\n",
                     )
-                    .short("m")
+                    .short('m')
                     .long("max")
                     .takes_value(true)
                     .default_value("0")
                     .validator(|d| d.parse::<u32>().map(|_| ()).map_err(|e| e.to_string())),
             )
             .arg(
-                Arg::with_name("REDUCE")
+                Arg::new("REDUCE")
                     .help("Reduce the max cell count when a result is found")
                     .long_help(
                         "Reduce the max cell count when a result is found\n\
                          The new max cell count will be set to the cell count of \
                          the current result minus one.\n",
                     )
-                    .short("R")
+                    .short('R')
                     .long("reduce")
                     .takes_value(false),
             )
             .arg(
-                Arg::with_name("SUBPERIOD")
+                Arg::new("SUBPERIOD")
                     .help("Allow patterns with subperiod")
                     .long_help("Allow patterns whose fundamental period are smaller than the given period")
-                    .short("p")
+                    .short('p')
                     .long("subperiod")
                     .takes_value(false),
             )
             .arg(
-                Arg::with_name("SKIPSUBSYM")
+                Arg::new("SKIPSUBSYM")
                     .help("Skip patterns invariant under more transformations than the given symmetry")
                     .long_help(
                         "Skip patterns which are invariant under more transformations than \
@@ -261,12 +248,12 @@ impl Args {
                          In another word, skip patterns whose symmetry group properly contains \
                          the given symmetry group.\n",
                     )
-                    .short("S")
+                    .short('S')
                     .long("skip-subsym")
                     .takes_value(false),
             )
             .arg(
-                Arg::with_name("BACKJUMP")
+                Arg::new("BACKJUMP")
                     .help("(Experimental) Enable backjumping")
                     .long_help(
                         "(Experimental) Enable backjumping\n\
@@ -281,26 +268,26 @@ impl Args {
         {
             app = app
                 .arg(
-                    Arg::with_name("ALL")
+                    Arg::new("ALL")
                         .help("Prints all possible results instead of only the first one")
                         .long_help(
                             "Prints all possible results instead of only the first one\n\
                              Only useful when --no-tui is set.\n",
                         )
-                        .short("a")
+                        .short('a')
                         .long("all")
                         .requires("NOTUI"),
                 )
                 .arg(
-                    Arg::with_name("RESET")
+                    Arg::new("RESET")
                         .help("Resets the time when starting a new search")
                         .long("reset-time")
                         .conflicts_with("NOTUI"),
                 )
                 .arg(
-                    Arg::with_name("NOTUI")
+                    Arg::new("NOTUI")
                         .help("Starts searching immediately, without entering the TUI")
-                        .short("n")
+                        .short('n')
                         .long("no-tui"),
                 );
         }
@@ -308,58 +295,43 @@ impl Args {
         #[cfg(not(feature = "tui"))]
         {
             app = app.arg(
-                Arg::with_name("ALL")
+                Arg::new("ALL")
                     .help("Searches for all possible pattern")
                     .long_help("Searches for all possible pattern")
-                    .short("a")
+                    .short('a')
                     .long("all"),
             );
         }
 
-        let matches = app.get_matches_safe()?;
+        let matches = app.clone().try_get_matches()?;
 
         let config;
 
         if let Some(config_file) = matches.value_of("CONFIG") {
             let path = Path::new(config_file);
-            let file = File::open(path)
-                .map_err(|e| Error::with_description(&e.to_string(), ErrorKind::Io))?;
+            let file = File::open(path).map_err(|e| app.error(ErrorKind::Io, e))?;
             let mut reader = BufReader::new(file);
             match path.extension().and_then(|s| s.to_str()) {
                 Some("json") => {
                     config = serde_json::from_reader(reader).map_err(|e| {
-                        Error::with_description(
-                            &format!("Invalid config file: {}", e),
-                            ErrorKind::Io,
-                        )
+                        app.error(ErrorKind::Io, format!("Invalid config file: {}", e))
                     })?;
                 }
                 Some("yaml" | "yml") => {
                     config = serde_yaml::from_reader(reader).map_err(|e| {
-                        Error::with_description(
-                            &format!("Invalid config file: {}", e),
-                            ErrorKind::Io,
-                        )
+                        app.error(ErrorKind::Io, format!("Invalid config file: {}", e))
                     })?;
                 }
                 Some("toml") => {
                     let mut buf = Vec::new();
                     reader
                         .read_to_end(&mut buf)
-                        .map_err(|e| Error::with_description(&e.to_string(), ErrorKind::Io))?;
+                        .map_err(|e| app.error(ErrorKind::Io, e))?;
                     config = toml::from_slice(&buf).map_err(|e| {
-                        Error::with_description(
-                            &format!("Invalid config file: {}", e),
-                            ErrorKind::Io,
-                        )
+                        app.error(ErrorKind::Io, format!("Invalid config file: {}", e))
                     })?;
                 }
-                _ => {
-                    return Err(Error::with_description(
-                        "Unsupported config file format",
-                        ErrorKind::InvalidValue,
-                    ))
-                }
+                _ => return Err(app.error(ErrorKind::Io, "Unsupported config file format")),
             }
         } else {
             let width = matches.value_of("X").unwrap().parse().unwrap();
@@ -422,14 +394,14 @@ impl Args {
         #[cfg(feature = "tui")]
         let no_tui = matches.is_present("NOTUI");
 
-        let world = config.world().map_err(|e| {
-            Error::with_description(&format!("Invalid config: {}", e), ErrorKind::InvalidValue)
-        })?;
+        let world = config
+            .world()
+            .map_err(|e| app.error(ErrorKind::InvalidValue, format!("Invalid config: {}", e)))?;
 
         if world.is_gen_rule() && config.backjump {
-            return Err(Error::with_description(
-                "Backjumping is not yet supported for Generations rules.",
+            return Err(app.error(
                 ErrorKind::InvalidValue,
+                "Backjumping is not yet supported for Generations rules.",
             ));
         }